split repl_commands_helpers.rs from repl_commands.rs
