+++ debugger/src/debug_adapter/dap.rs
@@ -10,8 +10,10 @@ pub(crate) mod dap_types;
 /// It doesn't make sense to implement all gdb commands, and this implementation will focus on the ones that
 /// are most useful to users, and this list is expected to grow over time.
 pub(crate) mod repl_commands;
-/// Various enums and structs used by the [`repl_commands::ReplCommand`].
+/// Helper functions to validate and build command subsets from [`super::repl_commands`]].
+pub(crate) mod repl_commands_helpers;
+/// Various enums and structs used by the [`super::repl_commands::ReplCommand`].
 pub(crate) mod repl_types;
 /// The logic for handling the various MS DAP requests, implemented so that it can be used by both the
-/// [`adapter::DebugAdapter`] and the [`repl_commands::ReplCommand`].
+/// [`adapter::DebugAdapter`] and the [`super::repl_commands::ReplCommand`].
 pub(crate) mod request_helpers;
